+++ pytorch_lightning/trainer/trainer.py
@@ -129,6 +129,7 @@ class Trainer(
             auto_lr_find: Union[bool, str] = False,
             replace_sampler_ddp: bool = True,
             progress_bar_callback: Optional[Union[ProgressBarBase, bool]] = True,
+            auto_scale_batch_size: Optional[str] = None,
             amp_level: str = 'O1',  # backward compatible, todo: remove in v0.8.0
             default_save_path=None,  # backward compatible, todo: remove in v0.8.0
             gradient_clip=None,  # backward compatible, todo: remove in v0.8.0
@@ -193,7 +194,7 @@ class Trainer(
             show_progress_bar:
                 .. warning:: .. deprecated:: 0.7.2
 
-                        Set `progress_bar_refresh_rate` to postive integer to enable. Will remove 0.9.0.
+                        Set `progress_bar_refresh_rate` to positive integer to enable. Will remove 0.9.0.
 
             progress_bar_refresh_rate: How often to refresh progress bar (in steps). Value ``0`` disables progress bar.
                 Ignored when a custom callback is passed to :paramref:`~Trainer.callbacks`.
@@ -293,6 +294,12 @@ class Trainer(
 
             terminate_on_nan: If set to True, will terminate training (by raising a `ValueError`) at the
                 end of each training batch, if any of the parameters or the loss are NaN or +/-inf.
+
+            auto_scale_batch_size: If set to True, will `initially` run a batch size
+                finder trying to find the largest batch size that fits into memory.
+                The result will be stored in self.hparams.batch_size in the LightningModule.
+                Additionally, can be set to either `power` that estimates the batch size through
+                a power search or `binsearch` that estimates the batch size through a binary search.
         """
 
         # Init callbacks
@@ -368,6 +375,7 @@ class Trainer(
         self.reload_dataloaders_every_epoch = reload_dataloaders_every_epoch
 
         self.auto_lr_find = auto_lr_find
+        self.auto_scale_batch_size = auto_scale_batch_size
         self.replace_sampler_ddp = replace_sampler_ddp
 
         self.truncated_bptt_steps = truncated_bptt_steps
@@ -474,7 +482,7 @@ class Trainer(
             self.show_progress_bar = show_progress_bar
 
         self.progress_bar_refresh_rate = progress_bar_refresh_rate
-        self.progress_bar_callback = None
+        self.progress_bar_callback = progress_bar_callback
         self.configure_progress_bar()
 
         # logging
@@ -737,6 +745,10 @@ class Trainer(
         # only on proc 0 because no spawn has happened yet
         model.prepare_data()
 
+        # Run auto batch size scaling
+        if self.auto_scale_batch_size:
+            self.scale_batch_size(model, mode=self.auto_scale_batch_size)
+
         # Run learning rate finder:
         if self.auto_lr_find:
             self._run_lr_finder_internally(model)
@@ -894,7 +906,7 @@ class Trainer(
             return
 
         # check if we should run validation during training
-        self.disable_validation = not (self.is_overriden('validation_step') and self.val_percent_check > 0) \
+        self.disable_validation = not (self.is_overridden('validation_step') and self.val_percent_check > 0) \
             and not self.fast_dev_run
 
         # run tiny validation (if validation defined)
@@ -996,52 +1008,51 @@ class Trainer(
 
         """
         # Check training_step, train_dataloader, configure_optimizer methods
-        if not self.is_overriden('training_step', model):
+        if not self.is_overridden('training_step', model):
             raise MisconfigurationException(
                 'No `training_step()` method defined. Lightning `Trainer` expects as minimum a'
-                ' `training_step()`, `training_dataloader()` and `configure_optimizers()` to be defined.')
+                ' `training_step()`, `train_dataloader()` and `configure_optimizers()` to be defined.')
 
-        if not self.is_overriden('train_dataloader', model):
+        if not self.is_overridden('train_dataloader', model):
             raise MisconfigurationException(
                 'No `train_dataloader()` method defined. Lightning `Trainer` expects as minimum a'
-                ' `training_step()`, `training_dataloader()` and `configure_optimizers()` to be defined.')
+                ' `training_step()`, `train_dataloader()` and `configure_optimizers()` to be defined.')
 
-        if not self.is_overriden('configure_optimizers', model):
+        if not self.is_overridden('configure_optimizers', model):
             raise MisconfigurationException(
                 'No `configure_optimizers()` method defined. Lightning `Trainer` expects as minimum a'
-                ' `training_step()`, `training_dataloader()` and `configure_optimizers()` to be defined.')
+                ' `training_step()`, `train_dataloader()` and `configure_optimizers()` to be defined.')
 
         # Check val_dataloader, validation_step and validation_epoch_end
-        if self.is_overriden('val_dataloader', model):
-            if not self.is_overriden('validation_step', model):
+        if self.is_overridden('val_dataloader', model):
+            if not self.is_overridden('validation_step', model):
                 raise MisconfigurationException('You have passed in a `val_dataloader()`'
                                                 ' but have not defined `validation_step()`.')
             else:
-                if not self.is_overriden('validation_epoch_end', model):
+                if not self.is_overridden('validation_epoch_end', model):
                     rank_zero_warn(
                         'You have defined a `val_dataloader()` and have defined a `validation_step()`,'
                         ' you may also want to define `validation_epoch_end()` for accumulating stats.',
                         RuntimeWarning
                     )
         else:
-            if self.is_overriden('validation_step', model):
+            if self.is_overridden('validation_step', model):
                 raise MisconfigurationException('You have defined `validation_step()`,'
                                                 ' but have not passed in a val_dataloader().')
 
         # Check test_dataloader, test_step and test_epoch_end
-        if self.is_overriden('test_dataloader', model):
-            if not self.is_overriden('test_step', model):
+        if self.is_overridden('test_dataloader', model):
+            if not self.is_overridden('test_step', model):
                 raise MisconfigurationException('You have passed in a `test_dataloader()`'
                                                 ' but have not defined `test_step()`.')
             else:
-                if not self.is_overriden('test_epoch_end', model):
+                if not self.is_overridden('test_epoch_end', model):
                     rank_zero_warn(
                         'You have defined a `test_dataloader()` and have defined a `test_step()`, you may also want to'
                         ' define `test_epoch_end()` for accumulating stats.', RuntimeWarning
                     )
 
     def check_testing_model_configuration(self, model: LightningModule):
-
         has_test_step = self.is_overriden('test_step', model)
         has_test_epoch_end = self.is_overriden('test_epoch_end', model)
         gave_test_loader = self.is_overriden('test_dataloader', model)