+++ tests/callbacks/test_callbacks.py
@@ -1,4 +1,5 @@
 import pytest
+
 import tests.base.utils as tutils
 from pytorch_lightning import Callback
 from pytorch_lightning import Trainer, LightningModule
@@ -11,7 +12,7 @@ from pathlib import Path
 def test_trainer_callback_system(tmpdir):
     """Test the callback system."""
 
-    hparams = tutils.get_default_hparams()
+    hparams = EvalModelTemplate.get_default_hparams()
     model = EvalModelTemplate(hparams)
 
     def _check_args(trainer, pl_module):
@@ -209,7 +210,7 @@ def test_early_stopping_no_val_step(tmpdir):
             output.update({'my_train_metric': output['loss']})  # could be anything else
             return output
 
-    model = CurrentModel(tutils.get_default_hparams())
+    model = CurrentModel()
     model.validation_step = None
     model.val_dataloader = None
 
@@ -245,7 +246,7 @@ def test_pickling(tmpdir):
 def test_model_checkpoint_with_non_string_input(tmpdir, save_top_k):
     """ Test that None in checkpoint callback is valid and that chkp_path is set correctly """
     tutils.reset_seed()
-    model = EvalModelTemplate(tutils.get_default_hparams())
+    model = EvalModelTemplate()
 
     checkpoint = ModelCheckpoint(filepath=None, save_top_k=save_top_k)
 
@@ -267,7 +268,7 @@ def test_model_checkpoint_with_non_string_input(tmpdir, save_top_k):
 def test_model_checkpoint_path(tmpdir, logger_version, expected):
     """Test that "version_" prefix is only added when logger's version is an integer"""
     tutils.reset_seed()
-    model = EvalModelTemplate(tutils.get_default_hparams())
+    model = EvalModelTemplate()
     logger = TensorBoardLogger(str(tmpdir), version=logger_version)
 
     trainer = Trainer(
@@ -286,7 +287,7 @@ def test_lr_logger_single_lr(tmpdir):
     """ Test that learning rates are extracted and logged for single lr scheduler"""
     tutils.reset_seed()
 
-    model = EvalModelTemplate(tutils.get_default_hparams())
+    model = EvalModelTemplate()
     model.configure_optimizers = model.configure_optimizers__single_scheduler
 
     lr_logger = LearningRateLogger()
@@ -311,7 +312,7 @@ def test_lr_logger_multi_lrs(tmpdir):
     """ Test that learning rates are extracted and logged for multi lr schedulers """
     tutils.reset_seed()
 
-    model = EvalModelTemplate(tutils.get_default_hparams())
+    model = EvalModelTemplate()
     model.configure_optimizers = model.configure_optimizers__multiple_schedulers
 
     lr_logger = LearningRateLogger()
@@ -329,4 +330,28 @@ def test_lr_logger_multi_lrs(tmpdir):
     assert len(lr_logger.lrs) == len(trainer.lr_schedulers), \
         'Number of learning rates logged does not match number of lr schedulers'
     assert all([k in ['lr-Adam', 'lr-Adam-1'] for k in lr_logger.lrs.keys()]), \
+        'Names of learning rates not set correctly'
+
+
+def test_lr_logger_param_groups(tmpdir):
+    """ Test that learning rates are extracted and logged for single lr scheduler"""
+    tutils.reset_seed()
+
+    model = EvalModelTemplate()
+    model.configure_optimizers = model.configure_optimizers__param_groups
+
+    lr_logger = LearningRateLogger()
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=5,
+        val_percent_check=0.1,
+        train_percent_check=0.5,
+        callbacks=[lr_logger]
+    )
+    results = trainer.fit(model)
+
+    assert lr_logger.lrs, 'No learning rates logged'
+    assert len(lr_logger.lrs) == 2 * len(trainer.lr_schedulers), \
+        'Number of learning rates logged does not match number of param groups'
+    assert all([k in ['lr-Adam/pg1', 'lr-Adam/pg2'] for k in lr_logger.lrs.keys()]), \
         'Names of learning rates not set correctly'